requirements.txt
flask~=2.2.5
flask-cors
boto3~=1.33.13
python-dotenv~=0.21.1
pytest~=7.4.4
jmespath~=1.0.1
pip~=24.0
attrs~=24.2.0
botocore~=1.33.13
s3transfer~=0.8.2
MarkupSafe~=2.1.5
Werkzeug~=2.2.3
click~=8.1.7
jinja2~=3.1.4
itsdangerous~=2.1.2
pluggy~=1.2.0
exceptiongroup~=1.2.2
iniconfig~=2.0.0
urllib3~=1.26.20
python-dateutil~=2.9.0.post0
six~=1.17.0
packaging~=24.0
moviepy